+++ src/Defaults.php
@@ -26,18 +26,6 @@ class Defaults
     private $phpVersions = ['5.6', '7.0', '7.1', '7.2'];
 
     /**
-     * Available php files to test on appveyor.
-     *
-     * @var array
-     */
-    private $phpAppVeyorVersions = [
-        '5.6' => 'php-5.6.29-nts-Win32-VC11-x86.zip',
-        '7.0' => 'php-7.0.17-nts-Win32-VC14-x86.zip',
-        '7.1' => 'php-7.1.3-nts-Win32-VC14-x64.zip',
-        '7.2' => 'php-7.2.0-nts-Win32-VC14-x64.zip',
-    ];
-
-    /**
      * Php versions without a semver scheme.
      *
      * @var array
@@ -124,16 +112,6 @@ class Defaults
     public function getPhpVersions()
     {
         return $this->phpVersions;
-    }
-
-    /**
-     * Get the available php versions to test on AppVeyor.
-     *
-     * @return array
-     */
-    public function getPhpAppVeyorVersions()
-    {
-        return $this->phpAppVeyorVersions;
     }
 
     /**