+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+#### v1.17.0 `2018-01-??`
+- `Added`
+    - Added Composer script descriptions to improve Composer script documentation when runtime Composer is >= `1.6.0`. Done by [@raphaelstolt](https://github.com/raphaelstolt). See [#215](https://github.com/jonathantorres/construct/issues/215).
+
 #### v1.16.0 `2017-10-26`
 - `Added`
     - Added PHP version 7.2 to the available versions. Done by [@raphaelstolt](https://github.com/raphaelstolt).